fix issue with clearing bit while analog stick is moving in a circle
with AnalogToDigital enabled, if a user used their stick and went to
the upper left. The joybutton for up and left would be set. If the user
then when to the bottom right, the down and right would be set, but the
up and left would not be cleared as the `SetBitIf` function can only
set bits. It will not clear them if the condition is false.