add support with android (oculus quest)
Android (Native Quest 2) builds use IL2CPP on unity. This presents an
issue where having the error is given
```
IL2CPP does not support marshaling delegates that point to instance methods to native code. The method we're attempting to marshal is: SK.Libretro.EnvironmentHandler::EnvironmentCallback
```
The workaround is to having the functions be declared as `static` and
are looked up within a Dictionary for the class instance with 'Thread'
as it's index.

Android also has a limitation with where it can dynamically link in code
The *.so library file of a libretro core just can not be from any location
on the filesystem for 'security reasons'. The location must be within
a specific directory. In this case, that directory is with in a temp
location. Where `tempDirectory` in this case initialized to
`$"{GetAndroidPrivateAppDataPath()}/temp"` and `libretroDirectory is
initialized to `$"{Application.persistentDataPath}/libretro"`.
GetAndroidPrivateAppDataPath() is shown below.
```cs
private string GetAndroidPrivateAppDataPath()
{
    string path;
    using (AndroidJavaClass jc = new AndroidJavaClass("com.unity3d.player.UnityPlayer"))
    {
        using (AndroidJavaObject currentActivity = jc.GetStatic<AndroidJavaObject>("currentActivity"))
        {
            path = currentActivity.Call<AndroidJavaObject>("getFilesDir").Call<string>("getCanonicalPath");
        }
    }
    return path;
}
```